test/requirements.txt
APScheduler
bilibili-api-python
edge-tts
pynput
duckduckgo_search
pyvirtualcam
opencv-python
numpy
Pillow
simplejson
websocket-client==1.7.0
flask
flask_apscheduler
peft
crawler